+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.4.0
 * #10: Make stm32loader useful as a library.
 * #4: Bring back support for progress bar.
 * #11: Support page erase in extended (two-byte addressing) erase mode
+* #12: Allow to supply the serial port as an environment variable
 * Start using code linting and unit tests.
 
 v0.3.3